Update workflow for new augur clades version
This updates the workflow to use augur after augur PR
https://github.com/nextstrain/augur/pull/728 has been merged. It removes
two rules from the workflow.

In the situation where the ncov workflow is not updated but augur is,
the pipeline will proceed without error however the 'emerging_lineage'
branch label will be missing. (This is because the add_branch_labels.py
script is looking for the key 'clade_annotation' which is no longer
present.)

In the situation where the ncov workflow is updated but augur is,
rule.emerging_lineages will error due to unknown arguments.

The script add_branch_labels.py is no longer used, but not removed here,
as it contains logic to export spike mutations as branch labels
which may be useful at some point. If we do use this, it would be better
to produce an intermediate node-data JSON with a custom branch label
to avoid modifying the auspice JSON after export.